requirements.txt
fastapi
vaderSentiment
uvicorn
mysql-connector-python
pydantic
python-jose
passlib
sqlalchemy
python-dotenv
python-multipart
# --- ML dependencies ---
pandas
numpy
scikit-learn
joblib
rapidfuzz
# --- Anaytics dependencies ---
openpyxl
pytest
requests
httpx
# pin bcrypt to a compatible version to avoid passlib backend detection issues
# (passlib expects bcrypt >= 3.1.0)
bcrypt>=3.1.0